backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pdfplumber==0.11.4
python-multipart==0.0.20
apscheduler==3.10.4
httpx==0.28.1
anthropic==0.42.0
openai==1.82.0
python-telegram-bot==21.9
python-jobspy==1.1.82
playwright==1.49.1
beautifulsoup4==4.12.3
pydantic==2.10.4
linkedin-api==2.3.1
Jinja2>=3.1.0
